requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
coverage==6.2
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
Pillow==8.4.0
psycopg2-binary==2.9.2
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
